El player ja no es torna boig en acabarse el temps...

+++ Pang/ModuleScenePlay.cpp
@@ -92,7 +92,7 @@ update_status ModuleScenePlay::Update()
 {
 
 	timer--;
-	if (timer <= 0)
+	if (timer <= 0 && App->player->dead = false)
 	{
 		App->player->Kill(8);
 	}
